+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added: `property-case` rule.
 - Added: `function-name-case` rule.
 - Added: `selector-pseudo-element-case` rule.
+- Added: `value-keyword-case` rule.
 - Fixed: `block-no-empty` no longer delivers false positives for less syntax.
 
 # 5.4.0